Move RefractiveIndex into its own module
